Refactor offchain data; Add epoch to ValidatorSnapshot
